apps/api/pyproject.toml
[project]
name = "educachile-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.1",
    "pydantic-settings==2.3.0",
    "sqlalchemy==2.0.36",
    "alembic==1.14.0",
    "psycopg[binary]==3.3.3",
    "celery[redis]==5.4.0",
    "redis==5.0.6",
    "minio==7.2.7",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "httpx==0.27.0",
    "email-validator==2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.2",
    "pytest-asyncio==0.23.7",
    "ruff==0.4.8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "alembic*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
